User-Rating-API/requirements.txt
-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
certifi==2022.9.24; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==2.1.1; python_full_version >= '3.6.0'
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3; python_version >= '3.6'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==4.1.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0; python_version >= '3.7'
djangorestframework-xml==2.0.0
djoser==2.1.0
idna==3.4; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
oauthlib==3.2.2; python_version >= '3.6'
pycparser==2.21
pyjwt==2.6.0; python_version >= '3.7'
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1; python_version >= '3.7' and python_version < '4'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
social-auth-app-django==4.0.0
social-auth-core==4.3.0; python_version >= '3.6'
sqlparse==0.4.3; python_version >= '3.5'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
